this seems to work for now, what a shit show
Basically we add children types of hole, const, varint, varbool
then evaluate based on them
Does lead to a lot of ugly match things since im terrible at
programming.
the synth loop could be cleaned (all of it could)
Hopefully it is not too much slower.

The next step is writing a front-end to parse problems to dynamically
add to define_language!